refactor: #2 코드 파일분리
